+++ source/lib/builtin.cpp
@@ -9,7 +9,7 @@ void builtin_abs(struct ccml::thread_t &thread) {
   const ccml::value_t *v = thread.pop();
   assert(v);
   if (v->is_int()) {
-    const int64_t res = (v->v < 0) ? (-v->v) : v->v;
+    const int32_t res = (v->v < 0) ? (-v->v) : v->v;
     thread.push_int(res);
   } else {
     thread.raise_error(thread_error_t::e_bad_argument);
@@ -21,7 +21,7 @@ void builtin_max(struct ccml::thread_t &thread) {
   const ccml::value_t *b = thread.pop();
   assert(a && b);
   if (a->is_int() && b->is_int()) {
-    const int64_t res = (a->v > b->v) ? a->v : b->v;
+    const int32_t res = (a->v > b->v) ? a->v : b->v;
     thread.push_int(res);
   } else {
     thread.raise_error(thread_error_t::e_bad_argument);
@@ -33,7 +33,7 @@ void builtin_min(struct ccml::thread_t &thread) {
   const ccml::value_t *b = thread.pop();
   assert(a && b);
   if (a->is_int() && b->is_int()) {
-    const int64_t res = (a->v < b->v) ? a->v : b->v;
+    const int32_t res = (a->v < b->v) ? a->v : b->v;
     thread.push_int(res);
   } else {
     thread.raise_error(thread_error_t::e_bad_argument);
@@ -45,12 +45,12 @@ void builtin_len(struct ccml::thread_t &t) {
   assert(a);
   switch (a->type) {
   case val_type_array: {
-    const int64_t res = a->array_size_;
+    const int32_t res = a->array_size_;
     t.push_int(res);
   } break;
   case val_type_string: {
     assert(a->s);
-    const int64_t res = a->s->size();
+    const int32_t res = a->s->size();
     t.push_int(res);
   } break;
   default:
